Optimizari la inchiderea programului

+++ Consumers/image_displayer.py
@@ -10,6 +10,7 @@ class ImageDisplayer:
     def __init__(self, queue):
         self._queue: Queue = queue
         self._process: Process = None
+        self._alive = True
 
     def start(self):
         Configurations.LOGGER.warning("CLIENT: Starting Image Displayer...")
@@ -18,18 +19,26 @@ class ImageDisplayer:
 
     def _start_receiving(self):
         namedWindow(Configurations.WINDOW_NAME, WINDOW_NORMAL)
-        while True:
+        while self._alive:
             try:
-                image = ImageOperations.decode(self._queue.get())
-                self.show_image(image)
-            except:
+                image = ImageOperations.decode(self._queue.get(timeout=1))
+                self._show_image(image)
+            except Exception as ex:
                 pass
 
     def stop(self):
         Configurations.LOGGER.warning("CLIENT: Stopping Image Displayer...")
-        self._process.kill()
-        destroyAllWindows()
+        if self._process is not None and self._process.is_alive():
+            self._alive = False
+            self._process.kill()
+            time.sleep(0.5)
+            self._process.close()
+            destroyAllWindows()
 
-    def show_image(self, image):
+    @staticmethod
+    def _show_image(image):
         imshow(Configurations.WINDOW_NAME, image)
         waitKey(1)
+
+    def is_alive(self):
+        return self._process.is_alive()
